==> ATAC-seq/scripts/dataAnalysis/05_annotation_openChromatinRegions.Rmd <==
---
title: "Open chromatin regions"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output:
  html_document:
    keep_md: true
    fig_width: 5
    fig_height: 5
    fig_caption: yes
    code_folding: hide
    toc: true
    toc_float: 
      collapsed: false
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(csaw)
library(GenomicFeatures)
library(org.Mm.eg.db)
library(ggplot2)
library(ggpubr)
library(RColorBrewer)
library(UpSetR)

dir <- "/Users/ibarra01/OneDrive - CRUK Cambridge Institute/github/somitogenesis2020/"

th <- theme_bw() + 
  theme(axis.text.x = element_text(size=10), 
        axis.title.x = element_text(size=12), 
        axis.text.y = element_text(size=10), 
        axis.title.y = element_text(size=12), 
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        axis.line = element_line(colour = "black"), 
        panel.border = element_blank(), 
        plot.title = element_text(face="bold", hjust = 0.5, size=12))

gg_color_hue <- function(n) {
  hues = seq(15, 375, length = n + 1)
  hcl(h = hues, l = 65, c = 100)[1:n]
}
```

We have defined a set of open chromatin loci, some of which show changes in accessibility across somites and development. Now we characterise these open chromatin regions.

```{r data}
## metadata
meta <- read.table(paste0(dir, "ATAC-seq/data/metadata_ATACseq_GQ.tsv"), 
                   stringsAsFactors = FALSE, header = TRUE)
meta <- meta[meta$QCpass==1,]

## peak counts
peaks <- readRDS(paste0(dir, "ATAC-seq/results/03_peakCounts_csawMerged.Rds"))
peaks <- rowRanges(peaks)

## ENSEMBL annotation doesn't prefix 'chr' to chromosome names; remove from peaks
seqlevelsStyle(peaks) <- "NCBI"
names(peaks) <- paste0(seqnames(peaks), ":", start(peaks), "-", end(peaks))
```


### Annotation of open chromatin regions

The open chromatin regions are considered putative regulatory elements, since they allow the interaction of regulatory proteins with the underlying DNA. These proteins can in turn modify gene expression. The two major, and best understood, regulatory element types are promoters and enhancers. Promoters are straight-forward to identify because they lie within genes' transcription start sites (TSS), and the gene catalog is very well annotated in the mouse genome. Enhancers, on the other hand, are more challenging since they can be located anywhere in the genome, although they tend to be *close* to the genes they regulate. Furthermore, other elements exist, such as repressors and insulators, that are not well-characterised and cannot be differentiated from enhancers without additional data.

Thus, we can first separate promoters from everything else, by identifying peaks that lie in TSS regions, using Ensembl's gene annotation. We define a *promoter* as the interval ±200bp around the TSS. This strategy results in almost a fifth of all peaks annotated as promoters:

```{r genomicAnn, message=FALSE, warning=FALSE}
## annotate open chromatin regions
## we use csaw function detailRanges to annotate the peaks
## but we want to use specifically Ensembl's annotation, version 96, to keep compatibility with the RNA-seq results
## thus, create a transcriptDB object from Ensembl v96
mmusculusEnsembl.v96 <- makeTxDbFromBiomart(biomart = "ENSEMBL_MART_ENSEMBL",
                                            dataset = "mmusculus_gene_ensembl",
                                            host="apr2019.archive.ensembl.org")
mcols(peaks) <- detailRanges(peaks,
                             txdb = mmusculusEnsembl.v96,
                             orgdb = org.Mm.eg.db,
                             dist = 5000, # flanking distance
                             promoter = c(200, 200),
                             key.field = "ENSEMBL")
peaks$promoter <- ifelse(grepl(":P", peaks$overlap), 1, 0)

c(nPeaks = sum(peaks$promoter), 
  pct = round(length(peaks[grep(":P", peaks$overlap),])/length(peaks)*100, 2))
```

To better annotate the peaks that do not overlap TSSs we need additional data. We exploit available data from ENCODE3 [https://screen.encodeproject.org/], FANTOM5 [https://fantom.gsc.riken.jp/5/] and the VISTA enhancer browser [https://enhancer.lbl.gov/].

### Annotation with external data

The ENCODE3 project has combined data on chromatin accessibility from DNAse-seq with H3K4me3, H3K27ac and CTCF ChIP-seq tracks to define promoter and enhancer elements. Data comes from a variety of cell lines and tissues and is combined into a consensus set of candidate cis-Regulatory Elements (cCREs).

```{r encode}
encode <- read.table(paste0(dir, "ATAC-seq/data/externalData/mm10-ccREs.bed"))
encode <- GRanges(seqnames = substr(encode$V1,4,15),
                  IRanges(start = encode$V2, end = encode$V3),
                  name = encode$V5,
                  ann = encode$V6)

## find overlaps
overlap.encode <- as.data.frame(peaks[queryHits(findOverlaps(peaks, encode)),], row.names = NULL)
colnames(overlap.encode) <- paste0("peaks_", colnames(overlap.encode))
overlap.encode <- cbind(overlap.encode, 
                        as.data.frame(encode[subjectHits(findOverlaps(peaks, encode)),]))
overlap.encode$locus <- paste0(overlap.encode$peaks_seqnames, ":", 
                               overlap.encode$peaks_start, "-", 
                               overlap.encode$peaks_end)

## concatenate all ENCODE annotations for each peak
peaks.externalAnn <- as.data.frame(peaks)
peaks.externalAnn$PLS <- ifelse(row.names(peaks.externalAnn) %in% 
                                  overlap.encode[grep("PLS", overlap.encode$ann),]$locus, 1, 0)
peaks.externalAnn$pELS <- ifelse(row.names(peaks.externalAnn) %in% 
                                   overlap.encode[grep("pELS", overlap.encode$ann),]$locus, 1, 0)
peaks.externalAnn$dELS <- ifelse(row.names(peaks.externalAnn) %in% 
                                   overlap.encode[grep("dELS", overlap.encode$ann),]$locus, 1, 0)
peaks.externalAnn$CTCF_bound <- ifelse(row.names(peaks.externalAnn) %in%
                                         overlap.encode[grep("CTCF-bound", overlap.encode$ann),]$locus, 1, 0)
peaks.externalAnn$H3K4me3 <- ifelse(row.names(peaks.externalAnn) %in%
                                      overlap.encode[grep("DNase-H3K4me3", overlap.encode$ann),]$locus, 1, 0)
```

From the `r length(encode)` annotated cCREs in the mouse genome, `r length(unique(overlap.encode$name))` (`r round(length(unique(overlap.encode$name))/length(encode)*100,2)`%) overlap ATAC-seq peaks in the somites. Conversely, out of the `r length(peaks)` peaks identified in the somites, `r length(unique(overlap.encode$locus))` (`r round(length(unique(overlap.encode$locus))/length(peaks)*100, 2)`%) overlap at least one cCRE. This indicates overall congruence between datasets, but also highlights many regulatory elements that haven't been captured by the cell populations used in ENCODE3 and, perhaps, somite-specific elements. 

FANTOM5 provides enhancer annotations based on CAGE data; bidirectional transcription initiation events have been show to mark enhancer elements. 

```{r fantom}
## blocks correspond to merged regions of transcription initiation evens in the plus and minus strand
fantom <- read.table(paste0(dir, "ATAC-seq/data/externalData/F5.mm10.enhancers.bed"))
colnames(fantom) <- c("chr", "start", "end", "name", "score", "strand", 
                      "thickStart", "thickEnd", "col", 
                      "blockCount", "blockSizes", "blockStarts")
fantom <- GRanges(seqnames = substr(fantom$chr, 4, 15),
                  IRanges(start = fantom$start, end = fantom$end))

## find overlaps
overlap.fantom <- as.data.frame(peaks[queryHits(findOverlaps(peaks, fantom)),], row.names = NULL)
colnames(overlap.fantom) <- paste0("peaks_", colnames(overlap.fantom))
overlap.fantom <- cbind(overlap.fantom, 
                        as.data.frame(fantom[subjectHits(findOverlaps(peaks, fantom)),]))
overlap.fantom$locus <- paste0(overlap.fantom$peaks_seqnames, ":", 
                               overlap.fantom$peaks_start, "-", 
                               overlap.fantom$peaks_end)

peaks.externalAnn$FANTOM <- ifelse(row.names(peaks.externalAnn) %in% overlap.fantom$locus, 1, 0)
```

Out of the `r length(fantom)` annotated enhancers in the mouse genome, `r nrow(unique(overlap.fantom[,10:12]))` (`r round(nrow(unique(overlap.fantom[,10:12]))/length(fantom)*100, 2)`%) overlap somite ATAC-seq peaks. This represents `r round(length(unique(overlap.fantom$locus))/length(peaks)*100, 2)`% of the ATAC-seq peaks.

Finally, the VISTA enhancer browser contains a collection of human and mouse putative regulatory elements tested in mouse transgenic models to assess whether these sequences are able to drive expression of a reporter gene, and determine tissue-specific expression patterns. For the human sequences, the orthologous mouse sequence has been defined. 

```{r vista}
vista <- read.table(paste0(dir, "ATAC-seq/data/externalData/VISTAenhancers_mm10.tsv"), 
                    sep="\t", header = TRUE)
vista <- GRanges(seqnames = vista$chr_mm10,
                 IRanges(start = vista$start_mm10, end = vista$end_mm10),
                 ID = vista$ID,
                 genes = vista$bracketing_genes,
                 expression = vista$expression)

## find overlaps
overlap.vista <- as.data.frame(peaks[queryHits(findOverlaps(peaks, vista)),], row.names = NULL)
colnames(overlap.vista) <- paste0("peaks_", colnames(overlap.vista))
overlap.vista <- cbind(overlap.vista, 
                       as.data.frame(vista[subjectHits(findOverlaps(peaks, vista)),]))
overlap.vista$locus <- paste0(overlap.vista$peaks_seqnames, ":", 
                              overlap.vista$peaks_start, "-",
                              overlap.vista$peaks_end)

peaks.externalAnn$VISTA <- ifelse(row.names(peaks.externalAnn) %in% overlap.vista$locus, 1, 0)

## how many annotation categories are present for each peak
peaks.externalAnn$sum <- rowSums(peaks.externalAnn[,9:16])

## capture whether the enhancer was positive or negative in embryos
peaks.externalAnn[peaks.externalAnn$VISTA==1,]$VISTA <- as.character(overlap.vista[
  match(row.names(peaks.externalAnn[peaks.externalAnn$VISTA==1,]), overlap.vista$locus),]$expression)
```

There are many fewer of these VISTA enhancers, with only `r length(vista)` tested elements, of which `r length(unique(overlap.vista$ID))` (`r round(length(unique(overlap.vista$ID))/length(vista)*100, 2)`%) overlap ATAC-seq peaks. Many of the sequences tested are quite large, so `r length(unique(overlap.vista$locus))` peaks overlap VISTA enhancers. From these, `r sum(unique(overlap.vista[,c(15,17)])[,2]=="Positive")` VISTA enhancers have positive expression in the reporter assay.

```{r in_somite}
## retrieve VISTA enhancer with expression annotated in somites
vista.somite <- read.table(paste0(dir, "ATAC-seq/data/externalData/VISTAenhancers_expr_somite.tsv"),
                           stringsAsFactors = FALSE, sep="\t", fill=TRUE)
vista.somite <- subset(vista.somite, vista.somite$V1 == "Human" | vista.somite$V1 == "Mouse")

ids <- unlist(lapply(strsplit(vista.somite$V3, " "), '[[', 3))
ids <- paste0(ifelse(vista.somite$V1=="Human", "hs", "mm"), ids)

## save which peaks are enhancers with *proved* expression in somites
peaks_VISTAsomite <- overlap.vista[overlap.vista$ID %in% ids,]
## and add this information to the general table
peaks.externalAnn$VISTAsomite <- ifelse(row.names(peaks.externalAnn) %in% peaks_VISTAsomite$locus, 
                                        1, 0)
```

Furthermore, `r length(ids)` of the VISTA enhancers with positive expression have positive signal in the somites; `r length(unique(peaks_VISTAsomite$ID))` of these overlap `r length(unique(peaks_VISTAsomite$locus))` ATAC-seq peaks.


### Classification of ATAC-seq peaks {.tabset}

We classify the peaks based on their genomic context:

- Promoter: if the peak is within 200bp of an annotated TSS.
- Exonic: if the peak overlaps an annotated exon.
- Intronic: if the peak overlaps an annotated intron
- Proximal: if the peak is within 25kb of an annotated gene.
- Distal: if the peak is farther than 25kb from an annotated gene, but within 100kb.
- Intergenic: any other peaks.

Recent studies linking putative enhancers to their target genes have shown that most enhancers regulate genes that are located *close-by*. Gasperini et al. (Cell, 2019) found that the median distance between enhancers and target genes was 24.1kb, and the vast majority of pairs were within 100kb. Thus, peaks that are very far from any genes are less likely to be regulators of gene expression.

Apart from promoters, most peaks fall within genes, mostly in introns. Very few peaks are in gene desserts (*intergenic* class, which doesn't have a gene within 100kb), with the vast majority within 25kb from an annotated gene.


```{r classify}
## compute distance to nearest gene for all peaks
dist <- distanceToNearest(peaks, genes(mmusculusEnsembl.v96))

## annotate each peak
peaks.classes <- peaks.externalAnn[,c(1:4,6)]
peaks.classes$nearest_gene <- mcols(dist)$distance

# precedence: promoter -> exon -> intron
peaks.classes$promoter <- ifelse(grepl(":P", peaks.classes$overlap), 1, 0)
peaks.classes$exonic <- ifelse(peaks.classes$promoter == 0 & grepl(":E", peaks.classes$overlap), 1, 0)
peaks.classes$intronic <- ifelse(peaks.classes$promoter == 0 & peaks.classes$exonic == 0 & grepl(":I", peaks.classes$overlap), 1, 0)
peaks.classes$proximal <- ifelse(peaks.classes$overlap == "" & 
                                   peaks.classes$nearest_gene <= 25e3, 
                                 1, 0)
peaks.classes$distal <- ifelse(peaks.classes$nearest_gene > 25e3 & 
                                 peaks.classes$nearest_gene <= 100e3, 
                               1, 0)
peaks.classes$intergenic <- ifelse(peaks.classes$nearest_gene > 100e3, 1, 0)

## for peaks in promoters, recover the gene
peaks.classes$promoter_gene <- NA
promoters <- promoters(mmusculusEnsembl.v96,
                       upstream = 200, downstream = 200,
                       columns = "GENEID")
# assign gene name based on ensembl ID
gene_ann <- read.table(paste0(dir, "RNA-seq/data/Mus_musculus.GRCm38.96.ann"),
                       stringsAsFactors = FALSE, row.names = 1)
promoters$geneName <- gene_ann[unlist(promoters$GENEID),1]
tmp <- peaks.classes[peaks.classes$promoter==1,]
for(i in 1:nrow(tmp)){
  # if(i %% 1000 == 0){ print(i) }
  peaks.classes[row.names(tmp)[i], 'promoter_gene'] <- paste(
    unique(unlist(
      promoters[subjectHits(findOverlaps(peaks[row.names(tmp)[i]], promoters))]$geneName)),
    collapse = ";")
}

## summarise
df <- data.frame(colSums(peaks.classes[,7:12]))
colnames(df) <- "count"
df$class <- row.names(df)
df$class <- factor(df$class, levels=c("promoter", "exonic", "intronic", "proximal", "distal", "intergenic"))

ggplot(df, aes(class, count, fill=class)) + 
  geom_bar(stat="identity") +
  geom_text(aes(label=count)) +
  xlab("") + ylab("number of peaks") +
  th + theme(legend.position = "none")

df$pct <- round(df$count/length(peaks)*100, 2)
df[,c(2,1,3)]
```

Therefore, the vast majority of the regions of open chromatin we identify in the somites have regulatory potential. 

Some of these have supporting evidence from the external databases, which makes them stronger candidates.

#### Promoter peaks

Not surprisingly, there is very good agreement between our classified promoters and the ENCODE2 cCREs with promoter-like signatures (PLS). Of note is that a large number of promoter peaks are bound by CTCF. As expected, only a small number of promoters overlap with FANTOM enhancers.

```{r promoters}
cols <- gg_color_hue(n=6)

upset(peaks.externalAnn[row.names(peaks.classes[peaks.classes$promoter==1,]),10:16],
                    main.bar.color = cols[1], order.by = "freq", nintersects = 10)
```

#### Genic peaks

Peaks that overlap genes but not the TSS are instead overlapped by distal enhancer-like elements. 

- Exonic peaks overlap more often with proximal cCREs compared to intronic peaks, which are almost always associated to distal cCREs.
- Exonic peaks are more often bound by CTCF, compared to intronic peaks.
- Intronic peaks overlap FANTOM enhancers more often than exonic peaks. 

```{r exonic}
print("Exonic")
upset(peaks.externalAnn[row.names(peaks.classes[peaks.classes$exonic==1,]),10:16], 
                    main.bar.color = cols[2], order.by = "freq", nintersects = 10)
```

```{r intronic}
print("Intronic")
upset(peaks.externalAnn[row.names(peaks.classes[peaks.classes$intronic==1,]),10:16],
                    main.bar.color = cols[3], order.by = "freq", nintersects = 10)
```

#### Proximal peaks

For peaks that do not overlap genes but are within 25kb we see a large overlap with ENCODE's dELS elements, and many are also identified in the FANTOM5 data.

```{r proximal}
upset(peaks.externalAnn[row.names(peaks.classes[peaks.classes$proximal==1,]),10:16],
                    main.bar.color = cols[4], order.by = "freq", nintersects = 10)
```

####  Distal peaks

A similar picture is observed for distal peaks, which are those that are even farther from genes, but within 100kb.

```{r distal}
upset(peaks.externalAnn[row.names(peaks.classes[peaks.classes$distal==1,]),10:16],
                    main.bar.color = cols[5], order.by = "freq", nintersects = 10)
```

####  Intergenic peaks

And only a few of the peaks that are away from genes have overlaps with the external data.

```{r intergenic}
upset(peaks.externalAnn[row.names(peaks.classes[peaks.classes$intergenic==1,]),10:16],
                    main.bar.color = cols[6], order.by = "freq", nintersects = 10)
```

###

---

Overall, promoter peaks are very well represented in the external databases, with over 85% annotated in ENCODE3. In contrast, over half of all peaks falling close to genes are not captured in the external catalogues. This is a reflection of the incomplete nature of these compendia, since only a limited set of cell types and tissues has been considered. 

```{r overlaps}
encode_ccres <- row.names(peaks.externalAnn[peaks.externalAnn$PLS == 1 | 
                                              peaks.externalAnn$pELS == 1 |
                                            peaks.externalAnn$dELS == 1, ])
ctcf_only <- setdiff(row.names(peaks.externalAnn[peaks.externalAnn$CTCF_bound == 1,]), 
                     encode_ccres)
fantom <- row.names(peaks.externalAnn[peaks.externalAnn$FANTOM == 1,])
vista <- row.names(peaks.externalAnn[peaks.externalAnn$VISTA != 0,])

peaks.classes$in_encode <- ifelse(row.names(peaks.classes) %in% encode_ccres, 1, 0)
peaks.classes$in_fantom <- ifelse(row.names(peaks.classes) %in% fantom, 1, 0)
peaks.classes$in_vista <- ifelse(row.names(peaks.classes) %in% vista, 1, 0)
peaks.classes$CTCF_only <- ifelse(row.names(peaks.classes) %in% ctcf_only, 1, 0)


df <- data.frame(promoters = colSums(peaks.classes[peaks.classes$promoter == 1, 14:17]),
                 exonic = colSums(peaks.classes[peaks.classes$exonic == 1, 14:17]),
                 intronic = colSums(peaks.classes[peaks.classes$intronic == 1, 14:17]),
                 proximal = colSums(peaks.classes[peaks.classes$proximal == 1, 14:17]),
                 distal = colSums(peaks.classes[peaks.classes$distal == 1, 14:17]),
                 intergenic = colSums(peaks.classes[peaks.classes$intergenic == 1, 14:17]))

df
df2 <- round(t(t(df)/colSums(peaks.classes[,7:12]))*100,2)
colnames(df2) <- paste0(colnames(df2), "%")
df2
```

Nonetheless, putative enhancer elements with available experimental data supporting their role as enhancers will be very useful. 

```{r save}
write.table(peaks.classes, paste0(dir, "ATAC-seq/results/05_peaks_classAnnotation.tsv"), 
            quote = FALSE, sep="\t")
write.table(peaks.externalAnn, paste0(dir, "ATAC-seq/results/05_peaks_overlapExternalData.tsv"), 
            quote = FALSE, sep="\t")
```


```{r info}
sessionInfo()
```
